\name{calcBankfullWidthGeomean}
\alias{calcBankfullWidthGeomean}
\title{Calculate the mean of the bankfull width}
\usage{
  calcBankfullWidthGeomean(x, uid)
}
\arguments{
  \item{x}{a vector of bankfull width measurements}

  \item{uid}{a vector of site-visit indicators}
}
\description{
  \code{calcBankfullWidthGeomean} returns the mean of the
  bankfull width at each site
}

