\name{calculateBankCoverVar}
\alias{calculateBankCoverVar}
\title{Calculate bank cover variability metrics}
\usage{
  calculateBankCoverVar(uid, parameter, cover)
}
\description{
  \code{calculateBankCoverVar} calculates a group of fish
  cover metrics dealing 'with the variability in cover
  provided by bank features (overhanging or undercut
  banks).  The metrics are sdfc_ucb, sdfc_ohv idrucb,
  idrohv, iqrucb, iqrohv
}

