\name{calculateBoatLittoralSubstrateMetrics}
\alias{calculateBoatLittoralSubstrateMetrics}
\title{Calculate non-wadeable substrate metrics}
\usage{
  calculateBoatLittoralSubstrateMetrics(uid, parameter,
    result)
}
\arguments{
  \item{uid}{a vector of site-visit indicators}

  \item{parameter}{a vector of parameter codes}

  \item{size.class}{a vector size classes}
}
\description{
  \code{calculateBoatSubstrateMetrics} is used to calculate
  non-wadeable substrate metrics
}

