% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChannelHabitatProportions.r
\name{calculateChannelHabitatProportions}
\alias{calculateChannelHabitatProportions}
\title{Calculate channel habitat metrics}
\usage{
calculateChannelHabitatProportions(uid, habitat, is.wadeable)
}
\arguments{
\item{uid}{a vector of site-visit indicators}

\item{habitat}{a vector of channel habitat codes}

\item{is.wadeable}{a logical vector with \code{TRUE} for wadeable sites}
}
\description{
\code{calculateChannelHabitatProportions} is used to calculate the channel
habitat metrics
}

