% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChannelMorphology.R
\name{calculateChannelMorphology}
\alias{calculateChannelMorphology}
\title{Calculate channel morphology metrics}
\usage{
calculateChannelMorphology(uid, transect, wetwid, bankwid, bankhgt, incishgt,
  depth)
}
\description{
This function joins extra transects and runs the metrics contained in \code{calculateChannelMetrics}
and \code{calculateWettedWidthMetrics}.
}
