% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateFishIBI.R
\name{calculateFishIBI}
\alias{calculateFishIBI}
\alias{fishTraits}
\title{Calculates Fish IBI}
\usage{
calculateFishIBI(site, order, species, length, anomaly,
  join.traits.by = "common_name", return.subindices = F,
  return.raw.metrics = F, ODFW.version = F)

fishTraits()
}
\arguments{
\item{site}{vector of site-visit identifiers}

\item{order}{a numeric vector; stream order correspoding to the location}

\item{species}{a vector of taxa names; must correspond to name field returned by \link{fishTraits}}

\item{length}{a vector of fish lengths in centimeters}

\item{anomaly}{a logical vector, \code{TRUE} if an anomaly was observed on the fish}

\item{join.traits.by}{which field from the data.frame returned by \link{fishTraits} should be used to match the species}

\item{return.subindices}{logical, should the subindices be returned or just the IBI?}

\item{return.raw.metrics}{should raw metrics be returned rather than IBI values?}
}
\description{
Calculates a fish IBI based on Hughes et al "A process for developing and evaluating indices
of fish assemblage integrity"
}
\details{
The \code{fishTraits} (unexported, use nrsa:::fishTraits to view code) function returns a \code{data.frame} with trait data for the IBI.
}

