% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateHumanInfluence.r
\name{calculateHIWeightedSums}
\alias{calculateHIWeightedSums}
\title{Calculate weighted human influence metrics}
\usage{
calculateHIWeightedSums(proportions, weights = c(0, 0.6667, 1, 1.5))
}
\arguments{
\item{proportions}{a uid x parameter x category matrix of proportions}

\item{weights}{a weight vector for categories 0, P, C, and B.}
}
\description{
Calculate weighted human influence metrics
}

