% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{calculateLWDCountMetrics}
\alias{calculateLWDCountMetrics}
\title{Calculate LWD metrics}
\usage{
calculateLWDCountMetrics(uid, size.class, count, is.wadeable, epa.names = T)
}
\arguments{
\item{uid}{vector of site-visit indicators}

\item{size.class}{A vector of size class codes (e.g., WSDSL, DXDLL, ...)}

\item{count}{vector of counts of LWD pieces}

\item{is.wadeable}{logical describing whether the site is wadeable or not}

\item{epa.names}{should the metrics be in data.frame format with EPA's names or in contingency table style.
Defaults to EPA style.}
}
\description{
This function calculates LWD counts within each site, size class, and whether
it is within the bankfull area
}
