% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{calculateLWDCumulativeMetrics}
\alias{calculateLWDCumulativeMetrics}
\title{Calculate cumulative LWD metrics}
\usage{
calculateLWDCumulativeMetrics(uid, size.class, count, is.wadeable, lwdlength,
  lwdarea, epa.names = T)
}
\arguments{
\item{uid}{vector of site-visit indicators}

\item{size.class}{A vector of size class codes (e.g., WSDSL, DXDLL, ...)}

\item{count}{vector of counts of LWD pieces}

\item{is.wadeable}{logical describing whether the site is wadeable or not}

\item{lwdlength}{numeric, the total stream length over which LWD has been quantified at the site.}

\item{lwdarea}{numeric, the total area over which LWD has been quantified at the site.}

\item{epa.names}{should the metrics be in data.frame format with EPA's names or in contingency table style.
Defaults to EPA style.}
}
\value{
If the \code{epa.names} argument is \code{FALSE}, then the
function returns an array of cumulative counts within size classes. 
If \code{TRUE}, the function returns a metric data.frame with epa 
metric names.
}
\description{
This function calculates count and volume metrics by size class.
}
\details{
The \code{lwdlength} and \code{lwdarea} argument should be as 
calculated by \link{calculateLWDSiteLength} and \link{calculateLWDSiteArea}
functions, respectively.
}

