% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{calculateLWDVolumeMetrics}
\alias{calculateLWDVolumeMetrics}
\alias{calculateLWDVolumeMetrics2}
\title{Calculate LWD volume metrics}
\usage{
calculateLWDVolumeMetrics(uid, zone, lwdlength, lwdarea, count, size, volume)

calculateLWDVolumeMetrics2(uid, zone, lwdlength, lwdarea, count, size, volume)
}
\arguments{
\item{uid}{a vector of site-visit indicators}

\item{zone}{a vector of 'd', 'w', and 't' values that describe whether the count was
observed within the bankfull ('w'), outside it ('d') or it's a total count ('t').}

\item{lwdlength}{the total length of the sampled lwd plots per site}

\item{lwdarea}{the total area of the sampled lwd plots per site}

\item{count}{counts of lwd pieces}

\item{size}{the size class of the LWD pieces}

\item{volume}{the calculated volume of the LWD size class}
}
\description{
\code{calculateLWDVolumeMetrics} is used to calculate the volume metrics from the
lwd counts at a site.
}

