\name{calculateS_RP100}
\alias{calculateS_RP100}
\title{Not even sure what this modified metric is...}
\usage{
  calculateS_RP100(xslope, sddepth)
}
\arguments{
  \item{xslope}{mean site slope}

  \item{sddepth}{standard deviation of thalweg depth}
}
\value{
  the s_rp100 metric
}
\description{
  Wish I knew.
}

