% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSubstrateEmbed.r
\name{calculateSubstrateEmbed}
\alias{calculateSubstrateEmbed}
\title{Calculates Substrate Embeddedness metrics
\code{calculateSubstrateEmbed} calculates substrate embeddedness metrics:
N33, N55, VCEMBED, VEMBED, XCEMBED, XEMBED.}
\usage{
calculateSubstrateEmbed(uid, embed, is.center)
}
\arguments{
\item{uid}{a vector of site-visit indicators}

\item{embed}{numeric vector of embeddedness values (0 - 100).}

\item{is.center}{a logical vector describing whether the embeddedness
is from a center point}
}
\description{
Calculates Substrate Embeddedness metrics
\code{calculateSubstrateEmbed} calculates substrate embeddedness metrics:
N33, N55, VCEMBED, VEMBED, XCEMBED, XEMBED.
}
