% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChannelMorphology.R
\name{calculateThalwegDepthMetrics}
\alias{calculateThalwegDepthMetrics}
\title{Calculate thalweg depth metrics: xdepth, sddepth, n_d}
\usage{
calculateThalwegDepthMetrics(uid, is.wadeable, depth, units = c("m", "cm",
  "mixed"))
}
\arguments{
\item{uid}{a vector of site identifiers}

\item{is.wadeable}{a logical vector, TRUE for each depth that is from
a wadeable protocol site (versus a non-wadeable protocol).}

\item{depth}{a vector of thalweg depths}

\item{desired}{output units.  Can be one of \code{"m"}, \code{"cm"}, or \code{"mixed"}.}
}
\value{
a 'metric' data.frame
}
\description{
Calculates the thalweg depth metrics (xdepth, sddepth, n_d) form thalweg depth
data.
}
\details{
The EPA code assumes input data in meters, which this function also assumes.
However, the EPA code reports the metrics for wadeable sites in centimeters.
This code includes an argument \code{units}, which allows the user to specify
the desired output units.  The \code{mixed} option mimics the EPA behavior.
}
