\name{calculateThalwegRatios}
\alias{calculateThalwegRatios}
\title{Calculate thalweg width depth ratios}
\usage{
  calculateThalwegRatios(uid, wetwid, depth)
}
\arguments{
  \item{uid}{}

  \item{wetwid}{}

  \item{depth}{}
}
\value{
  a 'metric' data.frame
}
\description{
  Calculate metrics dealing with the product and ratio of
  the channel width and depth
}

