% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSubstrateMetrics.r
\name{calculateWadeSubstrateMetrics}
\alias{calculateWadeSubstrateMetrics}
\title{Calculate wadeable substrate metrics}
\usage{
calculateWadeSubstrateMetrics(uid, size.class)
}
\arguments{
\item{uid}{a vector of site-visit indicators}

\item{size.class}{a vector size classes}
}
\description{
\code{calculateWadeSubstrateMetrics} is used to calculate wadeable substrate metrics: 
d16,d50,d84,dgm,lsub2d16,lsub2d16inor,lsub2d25,lsub2d50,lsub2d50inor,lsub2d75,
lsub2d84,lsub2d84inor,lsub2dmm,lsub2dmm_nor,lsub2iqr,lsubd2sd,lsubd2sd_nor, 
lsubd_sd,lsubd_sd_nor,lsub_d16,lsub_d25,lsub_d50,lsub_d75,lsub_d84,lsub_dmm, 
lsub_dmm_nor,lsub_iqr,n,n_nor,pct_bdrk,pct_bigr,pct_bl,pct_cb,pct_fn,pct_gc, 
pct_gf,pct_hp,pct_org,pct_om,pct_ot,pct_rc,pct_rr,pct_rs,pct_sa,pct_safn,pct_sb,
pct_sfgf,pct_wd,pct_xb,sub2dmm_nor,subd2sd_nor,subd_sd_nor,sub_dmm_nor
}

