% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{count.notna}
\alias{count.notna}
\title{Returns the number of non-missing values in the vector,}
\usage{
count.notna(x)
}
\arguments{
\item{x}{vector to determine n}
}
\value{
n
}
\description{
Returns the number of non-misingused to determine sample size.
}
