% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateResidualPools.R
\name{cutDimensions}
\alias{cutDimensions}
\title{Cuts a sequence of dimensions to only those dimensions that represent pools.}
\usage{
cutDimensions(x, na.omit = T)
}
\arguments{
\item{x}{a n * 2 matrix with column 1 being the residual depth and column 2
the residual length}

\item{na.omit}{should NAs be trimmed from the sequence.  This is to remove
the NAs from the diagonal and upper triangle pool dimension array returned
by \link{calculatePoolDimensions}.  These represent thalweg points that
come before the base point and are NA by definition.}
}
\description{
The \link{calculatePoolDimensions} function calculates pool dimensions (depth
and length) for every possible base point and each thalweg point after the
base point. For base points that actually define pools, the pool ends when
the residual depth becomes negative (or the site ends).  This function trims
a sequence of pool dimensions so that only the points in the sequence that
represent pools are returned.  In addition, the length is calculated as the
difference between thalweg sampling points (i.e., only the length
attributable to that point, rather than the cumulative length of the pool up
that point).
}

