% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateResidualPools.R
\name{extractResidualPoolDimensions}
\alias{extractResidualPoolDimensions}
\title{Extracts the dimensions of residual pools from the array of dimensions of
possible pools}
\usage{
extractResidualPoolDimensions(dimensions, base.point, is.pool)
}
\arguments{
\item{dimensions}{an array returned by \link{calculatePoolDimensions}}

\item{base.point}{a vector as long as the thalweg profile with the base point
for each pool}

\item{is.pool}{logical vector, \code{TRUE} if the depth represents a pool.}
}
\description{
The \link{calculatePoolDimensions} function calculates the dimensions of
every possible pool while treating each successive depth as a potential base
point. This function extracts those dimensions that actually represent pools
from the array. It returns a 3 * n matrix of residual length, depth and area.
}

