\name{fixBoatableDepthUnits}
\alias{fixBoatableDepthUnits}
\title{Convert boatable depth units from m to cm}
\usage{
  fixBoatableDepthUnits(x, is.wadeable)
}
\arguments{
  \item{x}{a depth based metric measured in m}

  \item{x}{logical, \code{TRUE} if the site was wadeable}
}
\value{
  a vector of depth metrics in cm
}
\description{
  Convert boatable depth units from m to cm
}

