% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelativeBedStability.R
\name{forceSlopeGt0}
\alias{forceSlopeGt0}
\title{Make slope slightly positive}
\usage{
forceSlopeGt0(slope)
}
\arguments{
\item{slope}{a vector of slopes}
}
\value{
a vector of slopes with slopes < .01, replaced by 0.01
}
\description{
Make slopes slightly positive if they are less than 0.01.
}

