\name{getNumTransects}
\alias{getNumTransects}
\title{Number of observed transects}
\usage{
  getNumTransects(uid, transect)
}
\arguments{
  \item{uid}{a vector of site-visit indicators}

  \item{transect}{a vector of transect ids}
}
\description{
  \code{getNumTransects} counts the number of observed
  transects per site
}

