% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateResidualPools.R
\name{getPoolInformation}
\alias{getPoolInformation}
\title{Determines basic information about a sequence of pools}
\usage{
getPoolInformation(dimensions)
}
\arguments{
\item{x}{a lower triangular matrix of residual pool dimensions.}
}
\value{
a data.frame of pool information.
}
\description{
Identifies pool base points, gives ids to individual pools, and numbers
points within a pool. from a matrix of residual pool depths.
}
\details{
The function first identifies which point on the thalweg profile is in a
pool. For a logical vector, the function then identifies runs in the logical
vector. Then it identifies the first point in each run.  Then sets the non
pools to FALSE. Then identifies which points are still TRUE and substracts 1
to get the previous point.
}

