% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{gmean}
\alias{gmean}
\title{Calculate the geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{numeric vectr}
}
\value{
the geometric mean of the vector
}
\description{
Calculates a geometric mean for a numeric vector.
}

