% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{\%&\%}
\alias{\%&\%}
\title{An AND operator that returns T for NA & T}
\usage{
e1 \%&\% e2
}
\arguments{
\item{e1}{logical vector}

\item{e2}{logical vector}
}
\value{
logical vector
}
\description{
An AND operator that returns NA for NA & NA NA %&% NA and returns TRUE for 
T %&% NA, mimicking the behavior of
function(x) {if(all(is.na(x))) { NA } else {all(x, na.rm=TRUE) } }
}
\examples{
e <- expand.grid(e1 = c(NA, T, F), e2 = c(NA, T, F))
e$e1 \%&\% e$e2
}

