% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateFishIBI.R
\name{indexfun}
\alias{indexfun}
\title{Create a function that maps raw metric scores to subindex scores}
\usage{
indexfun(zero, ten)
}
\arguments{
\item{zero}{numeric vector. The raw metric score representing a 0 subindex score. Can be of length 2, in which case the
first value is for stream order 1 and the second is for larger stream orders.}

\item{ten}{numeric vector of length 1 or 2. It is the value of the raw metric score that represents a 10 subindex score.

Returns a function that accepts two arguments, raw metric score and stream order and
returns the subindex score for the mapping specified.}
}
\description{
Create a function that maps raw metric scores to subindex scores for the Fish IBI using linear
interpolation. The subindex scores must be on a 0 to 10 scale. Raw metric scores
may be counts or percentages. The interpolation function used depends on the stream order of the site.
}

