% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{iqr}
\alias{iqr}
\title{Interquartile range with default type = 2}
\usage{
iqr(x, method = 2)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the interquartile range
}
\description{
Returns interquartile range (75%-25% of population, assuming gaussian
distribution) of the specified data.  Quantile calculation type 2 matches
SAS to within 10^-6 100% of the time (for 2008 data); this is at odds with
the R documentation which specifies type 3 imitates SAS.  FWIW, type 5
differs from SAS only 7% of the time under the same conditions.
}

