% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meltMetrics.r
\name{meltMetrics}
\alias{fillMissingMetrics}
\alias{meltMetrics}
\title{Melts metric data}
\usage{
meltMetrics(...)

fillMissingMetrics(x, uids, metrics)
}
\arguments{
\item{...}{metric data either in data.frame or matrix form; assumes in uid x metric form.}
}
\description{
Melts metric data. If data.frame, then there must be a UID column.  Dimnames must be present for a matrix.
}
\examples{
d1 <- data.frame(uid = 1:9, metric = 'xcl', result = 1:9)
fillMissingMetrics(d1, uids = 1:10, metric = 'xcl')
}

