% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{modalvalue}
\alias{modalvalue}
\title{Return the value which represents the mode}
\usage{
modalvalue(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values of interest}

\item{na.rm}{if TRUE, will remove NA values from consideration}
}
\value{
the value of the mode
}
\description{
Jason rewrote on 2011-11-02 and changed the test to reflect that the mode
will always be the smallest value if there are ties.  One test value returned
7 for a case with a 7 and 2 with three values each.
}

