\name{nrsa-package}
\alias{nrsa-package}
\alias{nrsa}
\docType{package}
\title{
Calculate National Rivers and Streams Assessment metrics.
}
\description{
This package can be used to calculate EMAP style metrics collected according
to the National Rivers and Streams Assessment protocol.
}
\details{
\tabular{ll}{
Package: \tab nrsa\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-10-20\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
The mets* functions are the workhorse functions.  The val* functions can be used
to check data validity.  The package requires that the data be present in a specific
database format.  See authors for format.
}
\author{
Jason Law
and EPA Western Ecology Division staff and contractors:
Marlys Cappaert
Curt Seeliger
?

Maintainer: Jason Law <jason.law@portlandoregon.gov>
}
\references{
USEPA. 2007. National Rivers and Streams Assessment:  Field Operations Manual. 
EPA-841-B-07-009. U.S. Environmental Protection Agency, Washington, DC
}
\keyword{ package }

