% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsa.options.r
\name{nrsa.options}
\alias{nrsa.options}
\title{Set nrsa package options}
\usage{
nrsa.options(...)
}
\arguments{
\item{...}{package options in tag = value form}
}
\description{
Sets nrsa package options. The only current option is the \code{ProgressReports} option.
Setting this option to \code{TRUE} will result in the functions of the package issuing 
progress reports when they are complete.
}
