\name{progressReport}
\alias{progressReport}
\title{Print a progress report
This function prints a progress report.  The time is appended so that the
elapsed time between reports can be used to track calculation progress.}
\usage{
  progressReport(msg)
}
\arguments{
  \item{msg}{the message to be printed}
}
\description{
  Print a progress report This function prints a progress
  report.  The time is appended so that the elapsed time
  between reports can be used to track calculation
  progress.
}

