% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculatePWQI.R
\name{PWQI}
\alias{PWQI}
\title{Calculate the Portland Water Quality Index
Calculate the Portland Water Quality Index as used in the watershed report cards.
The function is vectorized. Each argument should be the same length, excepting pH which
may be length 1.}
\usage{
PWQI(watershed, Cu, DO, Ecoli, Hg, NH4, P, TSS, Cel, pH = 7.8)
}
\arguments{
\item{watershed}{name of the watershed}

\item{Cu}{dissolved copper concentration in ug / L}

\item{DO}{dissolved oxygen in mg / L}

\item{Ecoli}{E. coli concentration in 1 / 100 mL}

\item{Hg}{total mercury concentration in ug / L}

\item{NH4}{ammonia-nitrogen concentration in mg / L}

\item{P}{phosphorus concentration in mg / L}

\item{TSS}{total suspended solids concentration in mg / L}

\item{Cel}{temperature in Celsius}

\item{pH}{pH}
}
\value{
A data.frame of water quality index subindices and the overall index.
}
\description{
Calculate the Portland Water Quality Index
Calculate the Portland Water Quality Index as used in the watershed report cards.
The function is vectorized. Each argument should be the same length, excepting pH which
may be length 1.
}

