% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFishIBI.R
\name{rescaleMetrics}
\alias{rescaleMetrics}
\title{Map fish IBI raw metric scores to rescaled subindex scores}
\usage{
rescaleMetrics(x)
}
\arguments{
\item{x}{a dataframe of raw metric scores}
}
\value{
a dataframe identical to input, except that metrics with names matching \code{nrsa:::kMetrics}
are rescaled to 0 - 10.
}
\description{
Maps raw metric scores to subindex scores for the Fish IBI using linear
interpolation. The subindex scores must be on a 0 to 10 scale. Raw metric scores
may be counts or percentages. The interpolation function used depends on the stream order of the site.
}
