% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sharedSupport.r
\name{which.is.max}
\alias{which.is.max}
\title{Which value is the max}
\usage{
which.is.max(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the index of the maximum
}
\description{
From package nnet
}

