% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{WoodClassesData}
\alias{WoodClassesData}
\title{LWD class metadata}
\usage{
WoodClassesData()
}
\value{
A data frame with the metadata
}
\description{
A data frame of LWD size class metadata that is in the LWD metric calculations.
}

