% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{calculateBankfullWidthGeomean}
\alias{calculateBankfullWidthGeomean}
\title{Calculate the mean of the bankfull width}
\usage{
calculateBankfullWidthGeomean(x, uid)
}
\arguments{
\item{x}{a vector of bankfull width measurements}

\item{uid}{a vector of site-visit indicators}
}
\description{
\code{calculateBankfullWidthGeomean} returns the mean of the bankfull width at each site
}

