\name{calculateBoatNumberWettedWidth}
\alias{calculateBoatNumberWettedWidth}
\title{Calculate boat protocol number of observed wetted widths}
\usage{
  calculateBoatNumberWettedWidth(uid, wetwid)
}
\arguments{
  \item{uid}{a vector of site identifiers}

  \item{wetwid}{a vector of wetted width measurements}
}
\value{
  a 'metric' data.frame
}
\description{
  Calculate boat protocol number of observed wetted widths:
  n_w
}

