% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHumanInfluence.r
\name{calculateHumanInfluence}
\alias{calculateHumanInfluence}
\title{Calculate human influence metrics}
\usage{
calculateHumanInfluence(uid, parameter, result, na.rm = F)
}
\arguments{
\item{uid}{a vector that differentiates among sites or site visits. Metrics will be
calculated for each value of \code{uid}.}

\item{parameter}{a vector of parameter names for the 11 human influence parameters:
must be one of \code{"row"}, \code{"past"}, \code{"build"}, \code{"landfl"}, \code{"log"},
\code{"mine"}, \code{"park"}, \code{"pave"}, \code{"row"}, \code{"pipes"}, \code{"road"}, 
or \code{"wall"}.}

\item{result}{a vector of results \code{0} (not present), \code{P} (present > 10 m), 
\code{C} (present < 10 m), \code{B} (present on bank).}
}
\value{
returns a dataframe with three columns: uid, metric, result
}
\description{
This function calculates the human influence metrics.  The parameter names are
assumed to be in lower case. Missing data is not allowed and will result
in an error.
}
\examples{
hi <- expand.grid(uid = 1:1000, transect = letters[1:10], 
                 parameter = c("wall", "build", "pave", "road", "pipes", "landfl", "park", 
                               "row", "past", "log", "mine"))
hi$result <- sample(c('0', 'C', 'B', 'P'), size = 11 * 10 * 1000, replace = T)
calculateHumanInfluence(hi$uid, hi$parameter, hi$result)
}
