% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateChannelMorphology.R
\name{calculateWettedWidthMetrics}
\alias{calculateWettedWidthMetrics}
\title{Calculates wetted width metrics}
\usage{
calculateWettedWidthMetrics(uid, wetwid)
}
\arguments{
\item{uid}{a vector of site identifiers}

\item{wetwid}{a vector of wetted widths; uses the max of the extra transect pairs}
}
\value{
a 'metric' data.frame
}
\description{
Calculates wetted width metrics: xwidth, sdwidth, and n_w.
}
\details{
Uses the max wetted width of the transect - extra transect pair as returned 
by \link{joinExtraTransects}.
}
