% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertNamedVectorToMetricDF.R
\name{convertNamedVectorToMetricDF}
\alias{convertNamedVectorToMetricDF}
\alias{convertVectorsToMetricDF}
\title{Convert a named vector to a metric data.frame}
\usage{
convertNamedVectorToMetricDF(x)

convertVectorsToMetricDF(...)
}
\arguments{
\item{x}{a named vector}
}
\value{
a 'metric' data.frame
}
\description{
Converts a named vector to a metric data.frame.  The names should be the site
identifiers, and the deparsed name of the the argument \code{x}, becomes the
metric name.
}
\examples{
my_metric <- c('uid1' = 0.1, 'uid2' = 0.2, 'uid3' = 0.3)
convertNamedVectorToMetricDF(my_metric)
}

