% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{getNumTransects}
\alias{getNumTransects}
\title{Number of observed transects}
\usage{
getNumTransects(uid, transect)
}
\arguments{
\item{uid}{a vector of site-visit indicators}

\item{transect}{a vector of transect ids}
}
\description{
\code{getNumTransects} counts the number of observed transects per site
}

