% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{observedLWDTransects}
\alias{observedLWDTransects}
\title{Calculate number of LWD observed transects}
\usage{
observedLWDTransects(uid, transect, count, in.bankfull)
}
\arguments{
\item{uid}{vector of site-visit indicators}

\item{transect}{vector of transect names}

\item{count}{vector of counts of LWD pieces}

\item{in.bankfull}{logical vector; true if the lwd was within bankfull}
}
\description{
\code{observedLWDTransects} is used to calculate the number of transects that
were observed for the LWD data.
}

