% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLargeWoody.r
\name{WoodClassesData}
\alias{WoodClassesData}
\title{LWD class metadata}
\usage{
WoodClassesData()
}
\value{
A data frame with the metadata
}
\description{
A data frame of LWD size class metadata that is in the LWD metric calculations. The volume
of a piece of LWD is determined by the forumla:
Volume = pi * [0.5*(minDiam + (maxDiam - minDiam)/3)]^2 * [minLength + (maxLength - minLength)/3]
The size variable is used to calculate the cumulative volume and cumulative 
count metrics (i.e., c1d, c1d_msq, etc).
}

